# services/api-gateway/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP edge for the Fireplace microservices. This spec is a\ncode-generated SHAPE contract (OpenAPI 2.0), not a validator:\nruntime validation lives in Go (gin `binding` tags at the edge,\ndomain invariants in the downstream services). Conditional\nrules are described in prose on each field, never encoded as\nschema constraints. See docs/api-conventions.md.",
        "title": "Fireplace API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/plans/{id}/checklists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists non-archived checklist items for a plan, optionally\nfiltered by scope and/or type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checklists"
                ],
                "summary": "List checklist items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "daily",
                            "longterm"
                        ],
                        "type": "string",
                        "description": "Filter by scope",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "task",
                            "note"
                        ],
                        "type": "string",
                        "description": "Filter by type",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ChecklistListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a checklist item (task or note) under the given plan.\n`scope` and `type` are optional shape enums. Parent nesting is\nvalidated downstream by plan-service (a parent must be a\ntop-level item in the same plan — two-tier maximum); that rule\nis enforced in code, not in this schema.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checklists"
                ],
                "summary": "Create a checklist item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Checklist item to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.CreateChecklistReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ChecklistResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/plans/{id}/checklists/archived": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists archived checklist items for a plan.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checklists"
                ],
                "summary": "List archived checklist items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ChecklistListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/plans/{id}/checklists/upcoming": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists checklist items whose start date falls within the next week.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checklists"
                ],
                "summary": "List upcoming checklist items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ChecklistListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/plans/{id}/checklists/{checklist_id}/archive": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Archives the item (sets archived=true). The request body is\nignored — this endpoint always archives, mirroring the monolith.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checklists"
                ],
                "summary": "Archive a checklist item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Checklist item ID (UUID)",
                        "name": "checklist_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ChecklistResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/plans/{id}/checklists/{checklist_id}/dates": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets, clears, or leaves the item's start/due dates. Both fields\nare OPTIONAL and three-state (omit = leave, null = clear,\n\"YYYY-MM-DD\" = set).\n\nCONDITIONAL RULE (enforced downstream in plan-service, described\nhere in prose and intentionally NOT encoded in the schema): when\nboth dates are present, `startDate` must be on or before\n`dueDate`. The spec is a shape contract, not a validator — see\ndocs/api-conventions.md.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "checklists"
                ],
                "summary": "Update checklist item dates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Checklist item ID (UUID)",
                        "name": "checklist_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dates to set/clear",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.UpdateDatesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ChecklistResponse"
                        }
                    },
                    "400": {
                        "description": "e.g. startDate after dueDate (enforced downstream)",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_gateway_plan.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_gateway_plan.ChecklistListResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_gateway_plan.ChecklistResp"
                    }
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "internal_gateway_plan.ChecklistResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/internal_gateway_plan.ChecklistResp"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "internal_gateway_plan.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "not found"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
...
        "internal_gateway_plan.UpdateDatesReq": {
            "type": "object",
            "properties": {
                "dueDate": {
                    "description": "New due date \"YYYY-MM-DD\", null to clear, or omit to leave unchanged. Must\nbe on or after startDate when both are present (enforced downstream).",
                    "type": "string",
                    "format": "date",
                    "example": "2026-06-15"
                },
                "startDate": {
                    "description": "New start date \"YYYY-MM-DD\", null to clear, or omit to leave unchanged.",
                    "type": "string",
                    "format": "date",
                    "example": "2026-06-01"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT bearer token. Format: \"Bearer \u0026lt;token\u0026gt;\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}